setup pipeline

+++ scripts/stop_server.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-if (ps aux | grep myWebApp | grep -v grep > /dev/null)
-  then
-      systemctl stop myWebApp
+CHECK=$(ps aux | grep myWebApp | grep -v grep)
+if [ ""="$CHECK"]; 
+    then
+        echo "myWebApp is not running"
+    else
+        systemctl stop myWebApp
 fi
